Split out the matching code from sceNetAdhoc.cpp
