+++ Core/HLE/proAdhoc.h
@@ -1131,10 +1131,6 @@ int findFreeMatchingID();
 */
 SceNetAdhocMatchingContext * findMatchingContext(int id);
 
-/*
-* Notify Matching Event Handler
-*/
-void notifyMatchingHandler(SceNetAdhocMatchingContext * context, ThreadMessage * msg, void * opt, u32_le &bufAddr, u32_le &bufLen, u32_le * args);
 // Notifiy Adhocctl Handlers
 void notifyAdhocctlHandlers(u32 flag, u32 error);
