+++ Core/HLE/sceNetAdhoc.h
@@ -20,19 +20,17 @@
 #include <deque>
 #include "Core/HLE/proAdhoc.h"
 
+
 #ifdef _MSC_VER
 #pragma pack(push,1)
 #endif
-typedef struct MatchingArgs {
-	u32_le data[6]; // ContextID, EventID, bufAddr[ to MAC], OptLen, OptAddr[, EntryPoint]
-} PACK MatchingArgs;
-
 typedef struct SceNetAdhocDiscoverParam {
 	u32_le unknown1; // SleepMode? (ie. 0 on on Legend Of The Dragon, 1 on Dissidia 012)
 	char   groupName[ADHOCCTL_GROUPNAME_LEN];
 	u32_le unknown2; // size of something? (ie. 0x3c on Legend Of The Dragon, 0x14 on Dissidia 012) // Note: the param size is 0x14 may be it can contains extra data too?
 	u32_le result; // inited to 0?
 } PACK SceNetAdhocDiscoverParam;
+
 #ifdef _MSC_VER
 #pragma pack(pop)
 #endif
@@ -102,7 +100,6 @@ void __NetAdhocInit();
 void __NetAdhocShutdown();
 void __NetAdhocDoState(PointerWrap &p);
 void __UpdateAdhocctlHandlers(u32 flags, u32 error);
-void __UpdateMatchingHandler(const MatchingArgs &params);
 
 bool __NetAdhocConnected();
 
@@ -114,7 +111,7 @@ int sceNetAdhocctlJoin(u32 scanInfoAddr);
 int sceNetAdhocctlScan();
 int sceNetAdhocctlGetScanInfo(u32 sizeAddr, u32 bufAddr);
 
-int NetAdhocMatching_Term();
+
 int NetAdhocctl_Term();
 int NetAdhocctl_GetState();
 int NetAdhocctl_Create(const char* groupName);
@@ -140,3 +137,9 @@ extern u32_le dummyThreadCode[3];
 extern u32 matchingThreadHackAddr;
 extern u32_le matchingThreadCode[3];
 
+// Exposing those for the matching routines
+int sceNetAdhocPdpSend(int id, const char* mac, u32 port, void* data, int len, int timeout, int flag);
+int sceNetAdhocPdpRecv(int id, void* addr, void* port, void* buf, void* dataLength, u32 timeout, int flag);
+int sceNetAdhocPdpCreate(const char* mac, int port, int bufferSize, u32 flag);
+int NetAdhoc_SetSocketAlert(int id, s32_le flag);
+int NetAdhocPdp_Delete(int id, int unknown);
